Unit-Prefixed Bytes

Index

GPUInspector.bytes

GPUInspector.bytes

GPUInspector.change_base

GPUInspector.simplify

GPUInspector.B

GPUInspector.GB

GPUInspector.GiB

GPUInspector.KB

GPUInspector.KiB

GPUInspector.MB

GPUInspector.MiB

GPUInspector.TB

GPUInspector.TiB

GPUInspector.UnitPrefixedBytes

References

GPUInspector.B — Type

Bytes

GPUInspector.GB — Type

Gigabytes, i.e. 10^9 = 1000^3 bytes

GPUInspector.GiB — Type

Gibibytes, i.e. 2^30 = 1024^3 bytes

GPUInspector.KB — Type

Kilobytes, i.e. 10^3 = 1000 bytes

GPUInspector.KiB — Type

Kibibytes, i.e. 2^10 = 1024 bytes

GPUInspector.MB — Type

Megabytes, i.e. 10^6 = 1000^2 bytes

GPUInspector.MiB — Type

Mebibytes, i.e. 2^20 = 1024^2 bytes

GPUInspector.TB — Type

Terabytes, i.e. 10^12 = 1000^4 bytes

GPUInspector.TiB — Type

Tebibytes, i.e. 2^40 = 1024^4 bytes

GPUInspector.UnitPrefixedBytes — Type

Abstract type representing an amount of data, i.e. a certain number of bytes, with a unit prefix (also "metric prefix"). Examples include the SI prefixes, like KB, MB, and GB, but also the binary prefixes (ISO/IEC 80000), like KiB, MiB, and GiB.

See https://en.wikipedia.org/wiki/Binary_prefix for more information.

GPUInspector.bytes — Method

bytes(x::Number)

Returns an appropriate UnitPrefixedBytes object, representing the number of bytes.

Note: This function is type unstable by construction!

See simplify for what "appropriate" means here.

GPUInspector.bytes — Method

bytes(x::UnitPrefixedBytes)

Return the number of bytes (without prefix) as Float64.

GPUInspector.change_base — Method

Toggle between

Base 10, SI prefixes, i.e. factors of 1000

Base 2, ISO/IEC prefixes, i.e. factors of 1024

Example:

julia> change_base(KB(13))
~12.7 KiB

julia> change_base(KiB(13))
~13.31 KB

GPUInspector.simplify — Method

simplify(x::UnitPrefixedBytes[; base])

Given a UnitPrefixedBytes number x, finds a more appropriate UnitPrefixedBytes that represents the same number of bytes but with a smaller value.

The optional keyword argument base can be used to switch between base 2, i.e. ISO/IEC prefixes (default), and base 10. Allowed values are 2, 10, :SI, :ISO, and :IEC.

Note: This function is type unstable by construction!

Example:

julia> simplify(B(40_000_000))
~38.15 MiB

julia> simplify(B(40_000_000); base=10)
40.0 MB